Web_App/requirements.txt
Flask==2.0.1
requests==2.26.0
tensorflow-cpu
Werkzeug==2.0.1
joblib==1.0.1
Pillow==8.3.1
gunicorn==20.1.0
lz4==3.1.3
scikit-learn==0.24.2
pandas==1.3.1